# api/docs/swagger.yaml
definitions:
  library_service.Author:
    properties:
      biography:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  library_service.BookRes:
    properties:
      author:
        $ref: '#/definitions/library_service.Author'
      genre:
        $ref: '#/definitions/library_service.Genre'
      id:
        type: string
      summary:
        type: string
      title:
        type: string
    type: object
  library_service.BorrowedBooksRes:
    properties:
      books:
        items:
          $ref: '#/definitions/library_service.BookRes'
        type: array
    type: object
  library_service.BorrowerCreateReq:
    properties:
      book_id:
        type: string
      user_id:
        type: string
    type: object
  library_service.BorrowerGetAllRes:
    properties:
      Borrowers:
        items:
          $ref: '#/definitions/library_service.BorrowerRes'
        type: array
    type: object
  library_service.BorrowerRes:
    properties:
      book:
        $ref: '#/definitions/library_service.BookRes'
      borrow_date:
        type: string
      id:
        type: string
      return_date:
        type: string
      user:
        $ref: '#/definitions/library_service.User'
    type: object
  library_service.BorrowerUpdate:
    properties:
      book_id:
        type: string
      borrow_date:
        type: string
      return_date:
        type: string
      user_id:
        type: string
    type: object
  library_service.Genre:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  library_service.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /borrowers:
    get:
      consumes:
      - application/json
      description: Get all borrower items
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      - description: book_id
        in: query
        name: book_id
        type: string
      - description: borrow_date
        in: query
        name: borrow_date
        type: string
      - description: return_date
        in: query
        name: return_date
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.BorrowerGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all borrower items
      tags:
      - Borrower
    post:
      consumes:
      - application/json
      description: Create a new borrower item
      parameters:
      - description: Borrower data
        in: body
        name: borrower
        required: true
        schema:
          $ref: '#/definitions/library_service.BorrowerCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.BorrowerRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create borrower item
      tags:
      - Borrower
  /borrowers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an borrower item by ID
      parameters:
      - description: Borrower ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Borrower item deleted
          schema:
            type: string
        "400":
          description: Invalid menu item ID
          schema:
            type: string
        "404":
          description: Menu item not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete menu borrower
      tags:
      - Borrower
    get:
      consumes:
      - application/json
      description: Get an borrower item by ID
      parameters:
      - description: Borrower ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.BorrowerRes'
        "400":
          description: Invalid menu item ID
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get borrower item
      tags:
      - Borrower
    put:
      consumes:
      - application/json
      description: Update an existing borrower item
      parameters:
      - description: Borrower ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated borrower item data
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/library_service.BorrowerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.BorrowerRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Menu item not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update borrower item
      tags:
      - Borrower
  /borrowers/overdue_book:
    get:
      consumes:
      - application/json
      description: Get overdue book items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.BorrowerGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get overdue book items
      tags:
      - Borrower
...
  /users/{user_id}/borrowed_books:
    get:
      consumes:
      - application/json
      description: Get all borrowed books
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.BorrowedBooksRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all borrowed books
      tags:
      - User
  /users/{user_id}/borrowing_history:
    get:
      consumes:
      - application/json
      description: Get all borrowed books history
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.BorrowedBooksRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all borrowed books history
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"